cleanup util and deck_id
