+++ s1_edh16_scrape.py
@@ -9,7 +9,7 @@ from selenium.webdriver.chrome.options import Options
 from selenium.webdriver.support.ui import WebDriverWait
 from selenium.webdriver.support import expected_conditions as EC
 from selenium.common.exceptions import TimeoutException, NoSuchElementException, StaleElementReferenceException
-from dir_utils import create_timestamped_dir
+from utils import create_timestamped_dir, extract_deck_id
 
 #todo improve the user input experience - automatically prompt the user for the weblink, or read the weblink from a config file
 #todo update the code to upload all files into their own new directory - no need to faff around rearranging stuff
@@ -101,18 +101,11 @@ def scrape_edhtop16(url):
 
             url = deck_link['href'].strip() #you have to do this because they sometimes make a small typo and add a space!!
 
-            if "topdeck.gg" in url:
+            deck_id = extract_deck_id(url)
 
-                deck_id = None
-
-            # Skip if not a Moxfield URL
-            elif "moxfield.com" in url:
-                # Extract the deck ID from the URL
-                deck_id_match = re.search(r'moxfield\.com/decks/([^/]+)', url)
-                if not deck_id_match:
-                    continue
-
-                deck_id = deck_id_match.group(1)
+            # Skip if not a valid Moxfield URL
+            if "moxfield.com" in url and not deck_id:
+                continue
 
             # Create deck info dictionary with the required fields
             deck_info = {
@@ -191,10 +184,7 @@ def scrape_edhtop16(url):
         # Always close the driver
         driver.quit()
 
-def main():
-    # Use the URL directly
-    # url = "https://edhtop16.com/commander/The%20Gitrog%20Monster?timePeriod=ONE_MONTH&minEventSize=100"
-    url = "https://edhtop16.com/commander/Kraum%2C%20Ludevic's%20Opus%20%2F%20Tymna%20the%20Weaver?timePeriod=ONE_MONTH&minEventSize=100"
+def main(url):
 
     # Scrape the data
     data = scrape_edhtop16(url)
@@ -233,6 +223,3 @@ def main():
 
     print(f"Scraped {len(data['decks'])} decks for {data['commander']}")
     print(f"Data saved to {output_file}")
-
-if __name__ == "__main__":
-    main()
